Fix webapi integration

+++ Rille.uTorrent.Extensions.PostProcess/Program.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using NLog;
 using Rille.uTorrent.Extensions.PostProcess.Model;
 using Rille.uTorrent.Extensions.PostProcess.Services;
+using System.Threading;
 
 namespace Rille.uTorrent.Extensions.PostProcess
 {
@@ -16,8 +17,28 @@ namespace Rille.uTorrent.Extensions.PostProcess
         private static FileManager _fileManager = new FileManager(_config);
         private static Unpacker unpacker = new Unpacker(_config, _fileManager);
         private static int processedTorrentsCount = 0;
+        static Mutex mutex = new Mutex(false, "https://github.com/rille111/Rille.uTorrent.Extensions");
 
         static void Main(string[] args)
+        {
+            // Wait 5 seconds if contended – in case another instance
+            // of the program is in the process of shutting down.
+            if (!mutex.WaitOne(TimeSpan.FromSeconds(5), false))
+            {
+                _logger.Debug("- Another instance of the app is running. Bye!");
+                return;
+            }
+            try
+            {
+                Run(args);
+            }
+            finally
+            {
+                mutex.ReleaseMutex();
+            }
+        }
+
+        private static void Run(string[] args)
         {
             try
             {
@@ -31,17 +52,22 @@ namespace Rille.uTorrent.Extensions.PostProcess
                 _logger.Info($"LOOPING TORRENTS");
                 foreach (var torrent in torrents)
                 {
+                    // TODO: TESTING
+                    //if (torrent.Hash != "2D8043305E789EDBBA5537E569D3DF56A6E6E3E8")
+                    //  continue;
+
                     LogStartProcessTorrent(torrent);
                     HandleAlreadyProcessedTorrent(torrent);
                     HandleUnprocessedTorrent(torrent);
+
+                    if (processedTorrentsCount >= _config.MaxProcessTorrentsInBatch)
+                    {
+                        // Enough in this batch already. Exit.
+                        _logger.Info($"- Already Processed {processedTorrentsCount} in this batch which was the configured max, exiting..");
+                        break;
+                    }
                 }
                 _logger.Info($"FINISHED");
-
-                //if (ShouldProcessAllTorrents())
-                //    ProcessAllTorrents();
-
-                //if (ShouldProcessOneTorrent())
-                //    ProcessOneTorrent();
             }
             catch (Exception ex)
             {
@@ -52,13 +78,13 @@ namespace Rille.uTorrent.Extensions.PostProcess
 
         private static void HandleUnprocessedTorrent(Torrent torrent)
         {
-            if (!_torrentManager.HasTorrentBeenPostProcessed(torrent))
+            if (torrent.IsDownloaded && !_torrentManager.HasTorrentBeenProcessed(torrent))
             {
                 if (processedTorrentsCount >= _config.MaxProcessTorrentsInBatch)
                 {
                     // Enough in this batch already. Exit.
                     _logger.Info($"- Already Processed {processedTorrentsCount} in this batch which was the configured max, exiting..");
-                    WaitAndExit();
+                    return;
                 }
 
                 // Otherwise, lets unpack/process!
@@ -71,36 +97,37 @@ namespace Rille.uTorrent.Extensions.PostProcess
                 if (unpackedOk)
                 {
                     // Mark torrent as finished
-                    _torrentManager.MarkTorrentAsProcessFinished(torrent);
+                    _torrentManager.MarkTorrentAsProcessed(torrent);
+                    _logger.Info($"- Torrent processed.");
 
-                    _logger.Info($"- Torrent processing finished.");
-
-                    if (_torrentManager.HasTorrentBeenPostProcessed(torrent) && _torrentManager.HasTorrentGoalsBeenReached(torrent))
+                    if (_torrentManager.HasTorrentBeenProcessed(torrent) && _torrentManager.HasTorrentGoalsBeenReached(torrent))
                     {
                         // Delete if goals reached and torrent processed ok, if configured as such.
                         if (_config.DeleteFromTorrentsFolderWhenUnpacked)
+                        {
                             _torrentManager.DeleteTorrent(torrent);
+                            _logger.Info("- Deleting processed torrent, goals has been reached.");
+                        }
                     }
                 }
                 else
                 {
                     // Unpack error!! Quit!
-                    _logger.Error($"- Failed to process torrent!");
+                    _logger.Error($"- Failed to process torrent! Investigate logs.");
                     _torrentManager.MarkTorrentAsProcessFailed(torrent);
-                    WaitAndExit();
                 }
             }
         }
 
         private static void HandleAlreadyProcessedTorrent(Torrent torrent)
         {
-            if (_torrentManager.HasTorrentBeenPostProcessed(torrent))
+            if (_torrentManager.HasTorrentBeenProcessed(torrent))
             {
-                // Torrent has already been processed.
-                _logger.Info($"- Torrent has already been processed!");
-                if (_config.DeleteAlreadyProcessedTorrents)
+                _logger.Info($"- Torrent has already been processed. Do anything with it?");
+                if (_torrentManager.HasTorrentGoalsBeenReached(torrent) && _config.DeleteAlreadyProcessedTorrents)
                 {
-                    // Delete already processed torrents was true, so delete it.
+                    // Torrents goal reached, and configured to deleted finished torrents, so delete it.
+                    _logger.Info("- Deleting processed torrent, goals has been reached.");
                     _torrentManager.DeleteTorrent(torrent);
                 }
             }
@@ -108,12 +135,18 @@ namespace Rille.uTorrent.Extensions.PostProcess
 
         private static void LogStartProcessTorrent(Torrent torrent)
         {
-            _logger.Info($"Processing: {torrent.Name} -");
+            _logger.Info($"Torrent: {torrent.Name} -");
+            _logger.Info($"- TorrentStatus: {torrent.TorrentStatus}");
+            _logger.Info($"- ProcessingStatus: {torrent.ProcessingStatus}");
+            _logger.Info($"- IsDownloaded: {torrent.IsDownloaded}");
+            _logger.Info($"- HasTorrentGoalsBeenReached: {_torrentManager.HasTorrentGoalsBeenReached(torrent)}");
+            _logger.Debug($"- Ratio: {torrent.ActualSeedRatioPercent}");
             _logger.Debug($"- Path: {torrent.Path}");
             _logger.Debug($"- IsFolder: {torrent.IsFolder}");
             _logger.Debug($"- IsSingleFileAndArchive: {torrent.IsSingleFileAndArchive}");
             _logger.Debug($"- IsSingleFileButNotArchive: {torrent.IsSingleFileButNotArchive}");
             _logger.Debug($"- DoesFolderContainAnyArchive: {_fileManager.DoesFolderContainAnyArchive(torrent.Path)}");
+            
         }
 
         private static void CreateTorrentManager()
@@ -137,40 +170,6 @@ namespace Rille.uTorrent.Extensions.PostProcess
                 var errors = string.Join("\n", result.Errors.Select(p => p.PropertyName + ": " + p.ErrorMessage + " Value: " + p.AttemptedValue));
                 throw new InvalidProgramException("Invalid configuration!\n" + errors);
             }
-        }
-
-        private static void ProcessOneTorrent()
-        {
-            //TODO: Support argument containing hash!
-            var torrent = new Torrent("....", _config);
-            _logger.Info($"Torrent hash is: {torrent.Hash}. I will now load all torrents.");
-
-            var torrents = _torrentManager.GetTorrentList();
-            torrent = torrents.Single(p => p.Hash == torrent.Hash);
-            _logger.Info($"Found torrent. Name: {torrent.Name}, NumericStatus: {torrent.NumericStatus}, Status: {torrent.TorrentStatus}");
-
-            if (!_torrentManager.HasTorrentBeenPostProcessed(torrent))
-            {
-                // Execute PostProcess
-                // TODO: Create method and move logging here
-                _logger.Info("Post processes starting.");
-            }
-
-            if (_torrentManager.HasTorrentBeenPostProcessed(torrent) && _torrentManager.HasTorrentGoalsBeenReached(torrent))
-            {
-                // Delete
-                // TODO: Create method and move logging here
-                _logger.Info("Deleting torrent, goals has been reached.");
-            }
-            else
-            {
-                _logger.Warn($"Torrent has either not been post processed yet, or the goals haven't been reached. Not doing anything.");
-            }
-        }
-
-        static void UnhandledExceptionTrapper(object sender, UnhandledExceptionEventArgs e)
-        {
-
         }
 
         private static void WaitAndExit()
